+++ src/tests/tribol_timestep_vote.cpp
@@ -133,6 +133,179 @@ TEST_F( CommonPlaneTest, zero_velocity_small_gap )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
+
+   int test_mesh_update_err = 
+      this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
+                                         tribol::FRICTIONLESS, false, parameters );
+
+   EXPECT_EQ( test_mesh_update_err, 0 );
+   EXPECT_EQ( parameters.dt, dt );
+
+   tribol::finalize();
+}
+TEST_F( CommonPlaneTest, large_velocity_small_gap_no_api_call )
+{
+   this->m_mesh.mortarMeshId = 0;
+   this->m_mesh.nonmortarMeshId = 1;
+
+   int nMortarElems = 4;
+   int nElemsXM = nMortarElems;
+   int nElemsYM = nMortarElems;
+   int nElemsZM = nMortarElems;
+
+   int nNonmortarElems = 5; 
+   int nElemsXS = nNonmortarElems;
+   int nElemsYS = nNonmortarElems;
+   int nElemsZS = nNonmortarElems;
+
+   // mesh bounding box with 'small' (0.055) interpenetration gap
+   real x_min1 = 0.;
+   real y_min1 = 0.;
+   real z_min1 = 0.; 
+   real x_max1 = 1.;
+   real y_max1 = 1.;
+   real z_max1 = 1.005;
+
+   real x_min2 = 0.;
+   real y_min2 = 0.;
+   real z_min2 = 0.95;
+   real x_max2 = 1.;
+   real y_max2 = 1.;
+   real z_max2 = 2.;
+
+   // compute element thickness for each block
+   real element_thickness1 = (z_max1 - z_min1) / nElemsZM;
+   real element_thickness2 = (z_max2 - z_min2) / nElemsZS;
+
+   // setup mesh
+   this->m_mesh.setupContactMeshHex( nElemsXM, nElemsYM, nElemsZM,
+                                     x_min1, y_min1, z_min1,
+                                     x_max1, y_max1, z_max1,
+                                     nElemsXS, nElemsYS, nElemsZS,
+                                     x_min2, y_min2, z_min2,
+                                     x_max2, y_max2, z_max2,
+                                     0., 0. );
+
+   // specify dt and component velocities for each block.
+   // Large velocity in the z-direction will incite a change in the 
+   // timestep. This velocity is computed on the high side using the 
+   // hardcoded rule that one face cannot interpen the other 
+   // exceeding 30% of the other's element thickness.
+   real dt = 1.e-3;
+   real bulk_mod1 = 1.0;
+   real bulk_mod2 = 1.0;
+   real vel_factor = 100.;
+   real velX1 = 0.;
+   real velY1 = 0.;
+   real velZ1 = vel_factor * 0.3 * element_thickness1 / dt;
+   real velX2 = 0.;
+   real velY2 = 0.;
+   real velZ2 = vel_factor * 0.3 * element_thickness2 / dt;
+
+   this->m_mesh.allocateAndSetVelocities( m_mesh.mortarMeshId, velX1, velY1, velZ1 );
+   this->m_mesh.allocateAndSetVelocities( m_mesh.nonmortarMeshId,  velX2, velY2, -velZ2 ); 
+   
+   // allocate and set element thickness and bulk modulus
+   this->m_mesh.allocateAndSetElementThickness( m_mesh.mortarMeshId, element_thickness1 );
+   this->m_mesh.allocateAndSetBulkModulus( m_mesh.mortarMeshId, bulk_mod1 );
+   this->m_mesh.allocateAndSetElementThickness( m_mesh.nonmortarMeshId, element_thickness2 );
+   this->m_mesh.allocateAndSetBulkModulus( m_mesh.nonmortarMeshId, bulk_mod2 );
+
+   // call tribol setup and update
+   tribol::TestControlParameters parameters; 
+   parameters.penalty_ratio = true;
+   parameters.const_penalty = 0.75;
+   parameters.dt = dt;
+   parameters.enable_timestep_vote = false;
+
+   int test_mesh_update_err = 
+      this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
+                                         tribol::FRICTIONLESS, false, parameters );
+
+   EXPECT_EQ( test_mesh_update_err, 0 );
+
+   EXPECT_EQ( parameters.dt, dt );
+
+   tribol::finalize();
+}
+
+TEST_F( CommonPlaneTest, numerically_zero_velocity_small_gap )
+{
+   // this test is meant to test tolerancing in the timestep calculation 
+   // when numerically zero velocities are present. This test caught a 
+   // negative dt estimate bug that has since been fixed.
+   this->m_mesh.mortarMeshId = 0;
+   this->m_mesh.nonmortarMeshId = 1;
+
+   int nMortarElems = 4;
+   int nElemsXM = nMortarElems;
+   int nElemsYM = nMortarElems;
+   int nElemsZM = nMortarElems;
+
+   int nNonmortarElems = 5; 
+   int nElemsXS = nNonmortarElems;
+   int nElemsYS = nNonmortarElems;
+   int nElemsZS = nNonmortarElems;
+
+   // mesh bounding box with 'small' (0.055) interpenetration gap
+   real x_min1 = 0.;
+   real y_min1 = 0.;
+   real z_min1 = 0.; 
+   real x_max1 = 1.;
+   real y_max1 = 1.;
+   real z_max1 = 1.005;
+
+   real x_min2 = 0.;
+   real y_min2 = 0.;
+   real z_min2 = 0.95;
+   real x_max2 = 1.;
+   real y_max2 = 1.;
+   real z_max2 = 2.;
+
+   // compute element thickness for each block
+   real element_thickness1 = (z_max1 - z_min1) / nElemsZM;
+   real element_thickness2 = (z_max2 - z_min2) / nElemsZS;
+
+   // setup mesh
+   this->m_mesh.setupContactMeshHex( nElemsXM, nElemsYM, nElemsZM,
+                                     x_min1, y_min1, z_min1,
+                                     x_max1, y_max1, z_max1,
+                                     nElemsXS, nElemsYS, nElemsZS,
+                                     x_min2, y_min2, z_min2,
+                                     x_max2, y_max2, z_max2,
+                                     0., 0. );
+
+   // specify dt and component velocities for each block.
+   // Large velocity in the z-direction will incite a change in the 
+   // timestep. This velocity is computed on the high side using the 
+   // hardcoded rule that one face cannot interpen the other 
+   // exceeding 30% of the other's element thickness.
+   real dt = 1.e-3;
+   real bulk_mod1 = 1.0; // something simple
+   real bulk_mod2 = 1.0; 
+   real velX1 = 1.e-12;
+   real velY1 = 1.e-12;
+   real velZ1 = 1.e-12; 
+   real velX2 = -1.e-12;
+   real velY2 = -1.e-12;
+   real velZ2 = -1.e-12; 
+
+   this->m_mesh.allocateAndSetVelocities( m_mesh.mortarMeshId, velX1, velY1, velZ1 );
+   this->m_mesh.allocateAndSetVelocities( m_mesh.nonmortarMeshId,  velX2, velY2, -velZ2 ); 
+   
+   // allocate and set element thickness and bulk modulus
+   this->m_mesh.allocateAndSetElementThickness( m_mesh.mortarMeshId, element_thickness1 );
+   this->m_mesh.allocateAndSetBulkModulus( m_mesh.mortarMeshId, bulk_mod1 );
+   this->m_mesh.allocateAndSetElementThickness( m_mesh.nonmortarMeshId, element_thickness2 );
+   this->m_mesh.allocateAndSetBulkModulus( m_mesh.nonmortarMeshId, bulk_mod2 );
+
+   // call tribol setup and update
+   tribol::TestControlParameters parameters; 
+   parameters.penalty_ratio = true;
+   parameters.const_penalty = 0.75;
+   parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
@@ -216,6 +389,7 @@ TEST_F( CommonPlaneTest, zero_velocity_large_gap )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
@@ -302,6 +476,7 @@ TEST_F( CommonPlaneTest, large_velocity_small_gap )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
@@ -309,7 +484,7 @@ TEST_F( CommonPlaneTest, large_velocity_small_gap )
 
    EXPECT_EQ( test_mesh_update_err, 0 );
 
-   real dt_diff = std::abs(parameters.dt - 0.000747973);
+   real dt_diff = std::abs(parameters.dt - 0.000997297);
    real dt_tol = 1.e-8;
    EXPECT_LT( dt_diff, dt_tol );
 
@@ -389,6 +564,7 @@ TEST_F( CommonPlaneTest, large_velocity_large_gap )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
@@ -396,7 +572,7 @@ TEST_F( CommonPlaneTest, large_velocity_large_gap )
 
    EXPECT_EQ( test_mesh_update_err, 0 );
 
-   real dt_diff = std::abs(parameters.dt - 2.02381e-06);
+   real dt_diff = std::abs(parameters.dt - 2.69841e-06);
    real dt_tol = 1.e-8;
    EXPECT_LT( dt_diff, dt_tol );
 
@@ -476,6 +652,7 @@ TEST_F( CommonPlaneTest, separation_velocity_small_gap )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
@@ -562,6 +739,7 @@ TEST_F( CommonPlaneTest, large_velocity_separation )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
@@ -649,13 +827,14 @@ TEST_F( CommonPlaneTest, large_velocity_small_separation )
    parameters.penalty_ratio = true;
    parameters.const_penalty = 0.75;
    parameters.dt = dt;
+   parameters.enable_timestep_vote = true;
 
    int test_mesh_update_err = 
       this->m_mesh.tribolSetupAndUpdate( tribol::COMMON_PLANE, tribol::PENALTY, 
                                          tribol::FRICTIONLESS, false, parameters );
 
    EXPECT_EQ( test_mesh_update_err, 0 );
-   real dt_diff = std::abs(parameters.dt -0.000749999);
+   real dt_diff = std::abs(parameters.dt -0.00099999);
    real dt_tol = 1.e-8;
    EXPECT_LT( dt_diff, dt_tol );