+++ src/tribol/common/Parameters.hpp
@@ -493,6 +493,7 @@ struct parameters_t
   int vis_cycle_incr;            ///! Frequency for visualizations dumps
   VisType vis_type;              ///! Type of interface physics visualization output
   std::string output_directory;  ///! Output directory for visualization dumps
+  bool enable_timestep_vote;     ///! True if host-code desires the timestep vote to be calculated and returned
 
 private: